fix install_colemak

+++ extra/keyboard/install_colemak.sh
@@ -1,7 +1,14 @@
 #!/usr/bin/bash
 
-echo "Installing colemak international custom keyboard"
 
-sudo cp ./symbols/custom /usr/share/X11/xkb/symbols
-sudo mv /usr/share/X11/xkb/rules/evdev.xml /usr/share/X11/xkb/rules/evdev.xml.bak
-sudo cp ./rules/evdev.xml /usr/share/X11/xkb/rules
+echo "__________________________________________________"
+echo "Installing colemak international custom keyboard"
+echo "__________________________________________________"
+
+local_dir=$(dirname "$0")
+
+sudo cp ${local_dir}/symbols/custom /usr/share/X11/xkb/symbols
+if [ -f /usr/share/X11/xkb/rules/evdev.xml ]; then
+    sudo mv /usr/share/X11/xkb/rules/evdev.xml /usr/share/X11/xkb/rules/evdev.xml.bak
+fi
+sudo cp ${local_dir}/rules/evdev.xml /usr/share/X11/xkb/rules
